scripts/requirements.txt
scholarly>=1.7.11
PyYAML>=6.0